chore: remove ChildEntitiesTagGroup component
